+++ module/actor/actor.js
@@ -4,6 +4,7 @@ import { CalculatePokemonCapabilities, CalculateTrainerCapabilities} from "./cal
 import { CalculateSkills } from "./calculations/skills-calculator.js"; 
 import { CalcBaseStat, CalculateStatTotal } from "./calculations/stats-calculator.js";
 import { GetMonEffectiveness } from "./calculations/effectiveness-calculator.js";
+import { warn, debug } from '../ptu.js' 
 
 /**
  * Extend the base Actor entity by defining a custom roll data structure which is ideal for the Simple system.
@@ -20,7 +21,7 @@ export class PTUActor extends Actor {
     const actorData = this.data;
 
     if(parseInt(game.data.version.split('.')[1]) <= 6) {
-      console.warn("FVTT PTU | Using old prepare-data structure")
+      warn("Using old prepare-data structure")
       // Make separate methods for each Actor type (character, npc, etc.) to keep
       // things organized.
       if (actorData.type === 'character') this._prepareCharacterData(actorData);
@@ -30,7 +31,7 @@ export class PTUActor extends Actor {
 
   /** @override */
   async modifyTokenAttribute(attribute, value, isDelta=false, isBar=true) {
-    console.log("FVTT PTU | ", attribute, value, isDelta, isBar);
+    debug("Modifying Token Attribute",attribute, value, isDelta, isBar);
     
     const current = getProperty(this.data.data, attribute);
     if (isBar) {
@@ -61,7 +62,7 @@ export class PTUActor extends Actor {
             value = current.max;
           }
         }
-        console.log("FVTT PTU | Updating Character HP with args:", this, {oldValue: current.value, newValue: value, tempHp: temp })
+        debug("Updating Character HP with args:", this, {oldValue: current.value, newValue: value, tempHp: temp })
         return this.update({[`data.${attribute}.value`]: value, [`data.${attribute}.temp.value`]: temp.value, [`data.${attribute}.temp.max`]: temp.max});
       }
       else {
@@ -78,7 +79,7 @@ export class PTUActor extends Actor {
   }
 
   /** @override */
-  prepareBaseData() {
+  prepareDerivedData() {
     const actorData = this.data;
 
     // Make separate methods for each Actor type (character, npc, etc.) to keep
@@ -91,7 +92,7 @@ export class PTUActor extends Actor {
   static async create(data, options={}) {
     let actor = await super.create(data, options);
 
-    console.log(actor);
+    debug("Creating new actor with data:",actor);
     if(actor.data.type != "pokemon") return;
 
     let form = new game.ptu.PTUPokemonCharactermancer(actor, {"submitOnChange": true, "submitOnClose": true});
@@ -120,11 +121,6 @@ export class PTUActor extends Actor {
     let dexExpEnabled = "true" == game.settings.get("ptu", "useDexExp") ?? false;
 
     // Make modifications to data here. For example:
-    data.levelUpPoints = 0;
-    for (let [key, value] of Object.entries(data.stats)) {
-        value["total"] = value["value"] + value["mod"] + value["levelUp"];      
-        data.levelUpPoints -= value["levelUp"];
-    }
 
     for (let [key, skill] of Object.entries(data.skills)) {
       skill["rank"] = this._getRank(skill["value"]);  
@@ -136,6 +132,11 @@ export class PTUActor extends Actor {
     else {
       data.level.current = data.level.milestones + Math.trunc(data.level.miscexp/10) + 1 > 50 ? 50 : data.level.milestones + Math.trunc(data.level.miscexp/10) + 1; 
     }
+
+    data.levelUpPoints = data.level.current + data.modifiers.statPoints + 9;
+    var result = CalculateStatTotal(data.levelUpPoints, data.stats, actorData.items.find(x => x.name.toLowerCase().replace("[playtest]") == "twisted power") != null);
+    data.stats = result.stats;
+    data.levelUpPoints = result.levelUpPoints;
     
     data.health.total = 10 + (data.level.current * 2) + (data.stats.hp.total * 3);
     data.health.max = data.health.injuries > 0 ? Math.trunc(data.health.total*(1-((data.modifiers.hardened ? Math.min(data.health.injuries, 5) : data.health.injuries)/10))) : data.health.total;
@@ -148,7 +149,6 @@ export class PTUActor extends Actor {
     data.ap.total = 5 + Math.floor(data.level.current / 5);
 
     data.initiative = {value: data.stats.spd.total + data.modifiers.initiative};
-    data.levelUpPoints += data.level.current + data.modifiers.statPoints + 9;
   }
 
   /**