+++ module/actor/character-sheet-gen8.js
@@ -1,4 +1,4 @@
-import { PrepareMoveData } from '../ptu.js'
+import { debug, error, log, PrepareMoveData } from '../ptu.js'
 
 /**
  * Extend the basic ActorSheet with some very simple modifications
@@ -165,7 +165,7 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 		delete itemData.data['type'];
 
 		// Finally, create the item!
-		console.log(itemData);
+		debug("Created new item",itemData);
 		return this.actor.createOwnedItem(itemData);
 	}
 
@@ -297,7 +297,7 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 				}),
 				move: move.data,
 				templateType: MoveMessageTypes.DETAILS
-			}).then(data => console.log(data))
+			}).then(data => debug(data))
 			return;
 		}
 		if(event.altKey) {
@@ -324,7 +324,7 @@ export class PTUGen8CharacterSheet extends ActorSheet {
 						}),
 						move: move.data,
 						templateType: MoveMessageTypes.DETAILS
-					}).then(data => console.log(data))
+					}).then(data => debug(data))
 				}
 			},
 			default: "roll"
@@ -378,7 +378,7 @@ function GetDiceResult(roll) {
 	try {
 		diceResult = roll.terms[0].results[0].result;
 	} catch (err) {
-		console.log("Old system detected, using deprecated rolling...")
+		log("Old system detected, using deprecated rolling...")
 		diceResult = roll.parts[0].results[0];
 	}
 	return diceResult;
@@ -391,7 +391,7 @@ function PerformAcRoll(roll, move, actor) {
 		}),
 		move: move.data,
 		templateType: MoveMessageTypes.TO_HIT
-	}).then(_ => console.log(`Rolling to hit for ${actor.name}'s ${move.name}`));
+	}).then(_ => log(`Rolling to hit for ${actor.name}'s ${move.name}`));
 
 	return GetDiceResult(roll);
 }
@@ -409,7 +409,7 @@ async function sendMoveRollMessage(rollData, messageData = {}) {
 	messageData.roll = rollData;
 
 	if(!messageData.move) {
-		console.error("Can't display move chat message without move data.")
+		error("Can't display move chat message without move data.")
 		return;
 	}
 	
@@ -426,7 +426,7 @@ async function sendMoveMessage(messageData = {}) {
 	}, messageData);
 
 	if(!messageData.move) {
-		console.error("Can't display move chat message without move data.")
+		error("Can't display move chat message without move data.")
 		return;
 	}