+++ module/custom-species.js
@@ -1,19 +1,20 @@
 import CustomSpeciesFolder from './entities/custom-species-folder.js'
+import {log, warn, error, debug} from './ptu.js'
 
 export async function InitCustomSpecies() {
     try {
         await CustomSpeciesFolder.initializeJournals();
-        CustomSpeciesFolder.updateFolderDisplay(game.settings.get("ptu", "hideDebugInfo"));
+        CustomSpeciesFolder.updateFolderDisplay(game.settings.get("ptu", "showDebugInfo"));
         
         window.PTUDebugger = {CustomSpeciesFolder};
 
         await migrateOldData();
 
-        console.log("FVTT PTU | Finalizing Custom Species Initialization")
+        log("Finalizing Custom Species Initialization")
         await Hooks.callAll("updatedCustomSpecies");
         await game.socket.emit("system.ptu", "RefreshCustomSpecies")
     } catch(err) {
-        console.warn("FVTT PTU | Unable to import data as player", err)
+        warn("Unable to import data as player", err)
         await ui.notifications.notify("There was a system update that requires the GM to login for migrations to happen. Please notify your GM and refresh after they have joined.", "warning")
     }
 }
@@ -29,7 +30,7 @@ async function migrateOldData() {
                     let journalEntry = CustomSpeciesFolder.findEntry(mon._id)
                     if(journalEntry) continue;
                     
-                    console.log("FVTT PTU | No entry found for " + mon._id + " creating new entry");
+                    log("No entry found for " + mon._id + " creating new entry");
                     let id = CustomSpeciesFolder.getAvailableId();
                     mergeObject(mon, {number: `${id}`, ptuNumber: id})
                     await JournalEntry.create({name: mon.ptuNumber, content: JSON.stringify(mon), folder: CustomSpeciesFolder._dirId})
@@ -38,12 +39,12 @@ async function migrateOldData() {
                 ui.notifications.notify("Imported old custom species data");
                 game.settings.set("ptu", "customSpecies", undefined);
             } catch(err) {
-                console.error("FVTT PTU | Unable to migrate old data", err)
+                error("Unable to migrate old data", err)
                 Dialog.confirm({
                     title: "Custom Species Data Migration",
                     content: "<p>For some reason we're unable to migrate your old species data. The error has been logged to the console</p><p>Would you like us to forget your old data?</p>",
                     yes: () => {
-                        console.log("FVTT PTU | Forgetting old custom species data.");
+                        log("Forgetting old custom species data.");
                         game.settings.set("ptu", "customSpecies", undefined);
                     },
                     no: () => {},
@@ -55,7 +56,8 @@ async function migrateOldData() {
 }
 
 export async function UpdateCustomSpecies(data) {
-    console.log("FVTT PTU | Triggering Custom Species Refresh with arguments:", data)
+    log("Triggering Custom Species Refresh")
+    debug("Custom Species Refresh Arguments:", data)
     game.ptu.customSpeciesData = await JSON.parse("["+Folders.instance.get(CustomSpeciesFolder._dirId).entities.map(x => $(`<p>${x.data.content}</p>`).text()).join(",")+"]");
     
     try {
@@ -66,13 +68,11 @@ export async function UpdateCustomSpecies(data) {
         }
     }
     catch(err) {
-        console.warn("FVTT PTU | Unable to update applications", err)
+        warn("Unable to update applications", err)
     }
 
     await game.actors.filter(x => x.data.type === "pokemon" && (x.data.data.isCustomSpecies || x.data.data.typing === undefined)).forEach(async (x) => {
         if(x.permission >= 3)
             await x.update({timestamp: Date.now()})
-        else 
-            await game.socket.emit("system.ptu", "ReloadGMSpecies")
     })
 }